requirements.txt
requests
matplotlib.pyplot
seaborn
datetime
streamlit
pandas